Criei um componente separado para a emissão de relatórios
